# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ReLink AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /relations/{relationId}/import/whatsapp:
    post:
      operationId: importWhatsapp
      tags: [whatsapp]
      summary: Import a WhatsApp export file
      parameters:
        - name: relationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WhatsappImportInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /relations/{relationId}/import/paste:
    post:
      operationId: importPaste
      tags: [whatsapp]
      summary: Import messages via copy-paste
      parameters:
        - name: relationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasteImportInput"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /relations/{relationId}/import/screenshot:
    post:
      operationId: importScreenshot
      tags: [whatsapp]
      summary: Analyze screenshot via OCR
      parameters:
        - name: relationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScreenshotImportInput"
      responses:
        "200":
          description: Extracted messages for validation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScreenshotExtractResult"
  /relations/{relationId}/import/screenshot/confirm:
    post:
      operationId: confirmScreenshotImport
      tags: [whatsapp]
      summary: Confirm and add screenshot messages
      parameters:
        - name: relationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScreenshotConfirmInput"
      responses:
        "200":
          description: Added messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
  /relations/{relationId}/import/manual:
    post:
      operationId: importManual
      tags: [whatsapp]
      summary: Add a message manually
      parameters:
        - name: relationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManualMessageInput"
      responses:
        "201":
          description: Added message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhatsappMessage"

  # ── WhatsApp Messages ──────────────────────────────────────────────────────
  /relations/{relationId}/messages:
    get:
      operationId: listMessages
      tags: [whatsapp]
      summary: List paginated messages
      parameters:
        - name: relationId
          in: path
          required: true
          schema:
            type: integer
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: search
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Paginated messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedMessages"
  /relations/{relationId}/messages/stats:
    get:
      operationId: getMessageStats
      tags: [whatsapp]
      summary: Get message statistics for a relation
      parameters:
        - name: relationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Message statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageStats"

  # ── Relational Memory ──────────────────────────────────────────────────────
...
components:
  schemas:
...
    WhatsappImportInput:
      type: object
      properties:
        content:
          type: string
          description: Raw content of the WhatsApp export file
        filename:
          type: string
      required: [content]
    PasteImportInput:
      type: object
      properties:
        text:
          type: string
        participantMe:
          type: string
        participantOther:
          type: string
      required: [text, participantMe, participantOther]
    ScreenshotImportInput:
      type: object
      properties:
        imageBase64:
          type: string
          description: Base64 encoded image
        mimeType:
          type: string
          default: "image/jpeg"
      required: [imageBase64]
    ScreenshotExtractResult:
      type: object
      properties:
        extractedMessages:
          type: array
          items:
            $ref: "#/components/schemas/ExtractedMessage"
        confidence:
          type: number
        needsReview:
          type: boolean
      required: [extractedMessages, confidence, needsReview]
    ExtractedMessage:
      type: object
      properties:
        tempId:
          type: string
        sender:
          type: string
        content:
          type: string
        timestamp:
          type: string
          nullable: true
        isMe:
          type: boolean
      required: [tempId, sender, content, isMe]
    ScreenshotConfirmInput:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ExtractedMessage"
        defaultDate:
          type: string
          format: date
      required: [messages]
    ManualMessageInput:
      type: object
      properties:
        sender:
          type: string
        content:
          type: string
        isMe:
          type: boolean
        sentAt:
          type: string
          format: date-time
      required: [sender, content, isMe, sentAt]
    WhatsappMessage:
      type: object
      properties:
        id:
          type: integer
        relationId:
          type: integer
        sender:
          type: string
        content:
          type: string
        isMe:
          type: boolean
        sentAt:
          type: string
          format: date-time
        importSource:
          type: string
          enum: [whatsapp_file, paste, screenshot, manual]
      required: [id, relationId, sender, content, isMe, sentAt, importSource]
    ImportResult:
      type: object
      properties:
        imported:
          type: integer
        duplicates:
          type: integer
        totalMessages:
          type: integer
        dateRange:
          type: object
          properties:
            from:
              type: string
              format: date
              nullable: true
            to:
              type: string
              format: date
              nullable: true
          required: [from, to]
      required: [imported, duplicates, totalMessages, dateRange]
    PaginatedMessages:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/WhatsappMessage"
        nextCursor:
          type: string
          nullable: true
        total:
          type: integer
      required: [messages, total]
    MessageStats:
      type: object
      properties:
        total:
          type: integer
        byParticipant:
          type: object
          additionalProperties:
            type: integer
        dateRange:
          type: object
          properties:
            from:
              type: string
              format: date
              nullable: true
            to:
              type: string
              format: date
              nullable: true
          required: [from, to]
        avgResponseTimeMinutes:
          type: number
          nullable: true
        daysWithMessages:
          type: integer
      required: [total, byParticipant, dateRange, daysWithMessages]

    # ── Memory ────────────────────────────────────────────────────────────────
...